bot/requirements.txt
python-telegram-bot==13.14
environs==9.5.0
requests==2.28.1